Drop miner hash constraint, import blocks together with addresses
